rename to prithvi for brevity
