+++ src/training/prithvi/train_hand.py
@@ -31,8 +31,8 @@ from terratorch.tasks import SemanticSegmentationTask
 project_root = Path(__file__).parent.parent.parent.parent
 sys.path.append(str(project_root))
 from src.data.io import IGNORE_INDEX
-from src.training.transformers_transfer_prithvi.eval import test_model
-from src.training.transformers_transfer_prithvi.sen1floods11 import (
+from src.training.prithvi.eval import test_model
+from src.training.prithvi.sen1floods11 import (
     Sen1Floods11S2HandDataModule,
 )
 
@@ -40,11 +40,7 @@ from src.training.transformers_transfer_prithvi.sen1floods11 import (
 ### SETTINGS
 SEED = 31415926
 DATA_PATH = Path(__file__).parent.parent.parent.parent / "data"
-OUTPUT_PATH = (
-    Path(__file__).parent.parent.parent.parent
-    / "outputs"
-    / "transformers_transfer_prithvi"
-)
+OUTPUT_PATH = Path(__file__).parent.parent.parent.parent / "outputs" / "prithvi"
 
 ALL_BAND_NAMES = (
     "COASTAL_AEROSOL",
